E-24: Authorization with Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -110,6 +110,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        // if ($thread->user_id != auth()->id()) {
+        //     abort(403, 'You do not permission to do this.');
+        // }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
